refactor: Separamos las vistas en una carpeta aparte para mejor la legibilidad y hacer mas simple el rest api
